+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - UPDATE: old Makefile removed
 - UPDATE: licence date update
 - UPDATE: test files updated
+- UPDATE: `dk cleani` removed please use `docker image prune` instead
 
 ## 17.06.0-ce (2017-07-05)
 - UPDATE: update to docker 17.06.0-ce